refactor: supabase api 경매 등록 메서드명  addAuction에서 insertAction으로 수정 (#118)
